Ajuste no tipo do campo da tabela de tarefas

+++ migrations/20210824020531_create_table_tasks.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
         table.string('descricao').notNull();
         table.integer('localizacao').notNull();
         table.integer('prioridade').notNull();
-        table.integer('entrega').notNull();
+        table.decimal('entrega').notNull();
         table.string('status').notNull();
         table.integer('noPrazo');
         table.integer('sequenciamento');
